+++ usvg/src/use_node.rs
@@ -106,12 +106,29 @@ pub(crate) fn convert_svg(
         new_ts.append(&ts);
     }
 
-    if let Some(clip_rect) = get_clip_rect(node, node, state) {
+    // We have to create a new state which would have its viewBox set to the current SVG element.
+    // Note that we're not updating State::size - it's a completely different property.
+    let mut state = state.clone();
+    state.view_box = {
+        if let Some(vb) = node.get_viewbox() {
+            vb
+        } else {
+            // No `viewBox` attribute? Then use `x`, `y`, `width` and `height` instead.
+            let x = node.convert_user_length(AId::X, &state, Length::zero());
+            let y = node.convert_user_length(AId::Y, &state, Length::zero());
+            let def = Length::new(100.0, LengthUnit::Percent);
+            let w = node.convert_user_length(AId::Width, &state, def);
+            let h = node.convert_user_length(AId::Height, &state, def);
+            Rect::new(x, y, w, h).unwrap_or(state.view_box)
+        }
+    };
+
+    if let Some(clip_rect) = get_clip_rect(node, node, &state) {
         let mut g = clip_element(node, clip_rect, orig_ts, id_generator, parent, tree);
-        convert_children(node, new_ts, state, id_generator, &mut g, tree);
+        convert_children(node, new_ts, &state, id_generator, &mut g, tree);
     } else {
         orig_ts.append(&new_ts);
-        convert_children(node, orig_ts, state, id_generator, parent, tree);
+        convert_children(node, orig_ts, &state, id_generator, parent, tree);
     }
 }